+++ packages/core/src/app/customer/StripeGuestForm.tsx
@@ -1,6 +1,7 @@
 import { CustomerInitializeOptions, CustomerRequestOptions } from '@bigcommerce/checkout-sdk';
 import { FieldProps, FormikProps, withFormik } from 'formik';
 import React, { FunctionComponent, memo, ReactNode, useCallback, useEffect, useState } from 'react';
+import { object } from 'yup';
 
 import { getAppliedStyles } from '@bigcommerce/checkout/dom-utils';
 import { TranslatedString, withLanguage, WithLanguageProps } from '@bigcommerce/checkout/locale';
@@ -46,7 +47,8 @@ const StripeGuestForm: FunctionComponent<StripeGuestFormProps & FormikProps<Gues
     requiresMarketingConsent,
     privacyPolicyUrl,
     step,
-    status
+    status,
+    setFieldValue,
 }) => {
 
     const [continueAsAGuestButton, setContinueAsAGuestButton] = useState(true);
@@ -61,6 +63,7 @@ const StripeGuestForm: FunctionComponent<StripeGuestFormProps & FormikProps<Gues
         });
     };
     const setEmailCallback = useCallback((authenticated: boolean, email: string) => {
+        setFieldValue('email', email);
         onChangeEmail(email);
         setEmailValue(email);
         setContinueAsAGuestButton(!email);
@@ -148,7 +151,6 @@ const StripeGuestForm: FunctionComponent<StripeGuestFormProps & FormikProps<Gues
             <div
                 className="optimizedCheckout-form-input"
                 id={ `${containerId}--input` }
-                placeholder="1111"
             >
                 <div
                     className="form-field--error"
@@ -270,6 +272,8 @@ export default withLanguage(
                             language,
                         })
                 }
+
+                return object({});
             },
         })(memo(StripeGuestForm))
 )