+++ packages/utility/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+    displayName: 'utility',
+    preset: '../../jest.preset.js',
+    transform: {
+        '^.+\\.[tj]sx?$': ['ts-jest', {
+            tsconfig: '<rootDir>/tsconfig.spec.json',
+            diagnostics: false,
+        }],
+    },
+    collectCoverageFrom: ['src/**/*.{ts,tsx}', '!src/**/*/index.ts'],
+    setupFilesAfterEnv: ['../../jest-setup.ts'],
+    coverageDirectory: '../../coverage/packages/utility',
+};